【功能完善】(package/yiapi): 完善menu/delete接口

+++ packages/yiapi/apis/menu/delete.js
@@ -1,11 +1,12 @@
+// 工具函数
 import { fnApiInfo } from '../../utils/index.js';
-
+// 配置文件
 import { appConfig } from '../../config/appConfig.js';
 import { codeConfig } from '../../config/codeConfig.js';
 import { metaConfig } from './_meta.js';
-
+// 接口信息
 const apiInfo = await fnApiInfo(import.meta.url);
-
+// 传参验证
 export const apiSchema = {
     tags: [apiInfo.parentDirName],
     summary: `删除${metaConfig.name}`,
@@ -18,30 +19,20 @@ export const apiSchema = {
         required: ['id']
     }
 };
-
+// 处理函数
 export default async function (fastify, opts) {
     fastify.post(`/${apiInfo.pureFileName}`, {
         schema: apiSchema,
         handler: async function (req, res) {
             try {
-                let menuModel = fastify.mysql //
-                    .table('sys_menu');
+                const menuModel = fastify.mysql.table('sys_menu');
 
-                let menuData = await menuModel.clone().where({ id: req.body.id }).first();
+                const menuData = await menuModel.clone().where({ id: req.body.id }).first('id');
 
-                if (!menuData) {
+                if (!menuData?.id) {
                     return {
                         ...codeConfig.DELETE_FAIL,
                         msg: '菜单不存在'
-                    };
-                }
-
-                let childData = await menuModel.clone().where({ pid: req.body.id }).first();
-
-                if (childData) {
-                    return {
-                        ...codeConfig.DELETE_FAIL,
-                        msg: '存在子菜单，无法删除'
                     };
                 }
 
@@ -52,7 +43,16 @@ export default async function (fastify, opts) {
                     };
                 }
 
-                let result = await menuModel.clone().where({ id: req.body.id }).delete();
+                const childData = await menuModel.clone().where({ pid: req.body.id }).first('id');
+
+                if (childData?.id) {
+                    return {
+                        ...codeConfig.DELETE_FAIL,
+                        msg: '存在子菜单，无法删除'
+                    };
+                }
+
+                const result = await menuModel.clone().where({ id: req.body.id }).delete();
                 await fastify.cacheTreeData();
                 return {
                     ...codeConfig.DELETE_SUCCESS,
